requirements.txt
asgiref==3.3.1
beautifulsoup4==4.9.3
boto3==1.17.53
botocore==1.20.53
certifi==2020.12.5
Django==3.1.7
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==2.3.1
django-crispy-forms==1.11.1
django-storages==1.11.1
django-urls==1.1.3
gunicorn==20.1.0
jmespath==0.10.0
psycopg2==2.8.6
python-dateutil==2.8.1
pytz==2021.1
s3transfer==0.3.7
six==1.15.0
soupsieve==2.2
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
wincertstore==0.2
